+++ test/functional/abc_rpc_getavalancheinfo.py
@@ -9,8 +9,10 @@ from test_framework.address import ADDRESS_ECREG_UNSPENDABLE
 from test_framework.avatools import (
     AvaP2PInterface,
     avalanche_proof_from_hex,
+    can_find_inv_in_poll,
     create_coinbase_stakes,
     gen_proof,
+    get_ava_p2p_interface,
     get_ava_p2p_interface_no_handshake,
     wait_for_proof,
 )
@@ -26,6 +28,8 @@ from test_framework.util import (
 )
 from test_framework.wallet_util import bytes_to_wif
 
+QUORUM_NODE_COUNT = 16
+
 # Interval between 2 proof cleanups
 AVALANCHE_CLEANUP_INTERVAL = 5 * 60
 # Dangling proof timeout
@@ -36,6 +40,7 @@ class GetAvalancheInfoTest(BitcoinTestFramework):
     def set_test_params(self):
         self.num_nodes = 1
         self.conflicting_proof_cooldown = 100
+        self.noban_tx_relay = True
         self.extra_args = [
             [
                 f"-avalancheconflictingproofcooldown={self.conflicting_proof_cooldown}",
@@ -757,9 +762,10 @@ class GetAvalancheInfoTest(BitcoinTestFramework):
             "Expire the local proof and check the verification status is now invalid"
         )
 
+        now = int(time.time())
         node.setmocktime(proof.expiration + 1)
         # Expiry is based on MTP, so we have to generate 6 blocks
-        self.generate(node, 6, sync_fun=self.no_op)
+        first_future_block = self.generate(node, 6, sync_fun=self.no_op)[0]
         # Check the proof status is what we expect
         assert_raises_rpc_error(
             -8,
@@ -778,6 +784,124 @@ class GetAvalancheInfoTest(BitcoinTestFramework):
             )
 
         self.wait_until(local_status_invalid)
+
+        self.log.info("Check the local proof status when it is avalanche invalidated")
+
+        # we need to revert the last 6 blocks that are in the future, or the
+        # node will not restart.
+        node.invalidateblock(first_future_block)
+        node.setmocktime(now + 1)
+
+        privkey, proof = gen_proof(
+            self,
+            node,
+        )
+        self.restart_node(
+            0,
+            self.extra_args[0]
+            + [
+                f"-avaproof={proof.serialize().hex()}",
+                f"-avamasterkey={bytes_to_wif(privkey.get_bytes())}",
+                "-avaproofstakeutxoconfirmations=1",
+                "-avalancheconflictingproofcooldown=0",
+                "-persistavapeers=0",
+            ],
+        )
+
+        quorum = [
+            get_ava_p2p_interface(self, node) for _ in range(0, QUORUM_NODE_COUNT)
+        ]
+        assert node.getavalancheinfo()["ready_to_poll"] is True
+
+        # Trigger validation by mining a block
+        self.generate(node, 1)
+
+        self.wait_until(
+            lambda: node.getavalancheinfo()
+            == {
+                "ready_to_poll": True,
+                "local": {
+                    "verified": True,
+                    "proofid": uint256_hex(proof.proofid),
+                    "limited_proofid": uint256_hex(proof.limited_proofid),
+                    "master": privkey.get_pubkey().get_bytes().hex(),
+                    "stake_amount": coinbase_amount,
+                    "payout_address": ADDRESS_ECREG_UNSPENDABLE,
+                },
+                "network": {
+                    "proof_count": QUORUM_NODE_COUNT + 1,
+                    "connected_proof_count": QUORUM_NODE_COUNT + 1,
+                    "dangling_proof_count": 0,
+                    "finalized_proof_count": 0,
+                    "conflicting_proof_count": 0,
+                    "immature_proof_count": 0,
+                    "total_stake_amount": (QUORUM_NODE_COUNT + 1) * coinbase_amount,
+                    "connected_stake_amount": (QUORUM_NODE_COUNT + 1) * coinbase_amount,
+                    "dangling_stake_amount": Decimal("0.00"),
+                    "immature_stake_amount": Decimal("0.00"),
+                    "node_count": QUORUM_NODE_COUNT + 1,
+                    "connected_node_count": QUORUM_NODE_COUNT + 1,
+                    "pending_node_count": 0,
+                },
+            }
+        )
+
+        def wait_for_invalidated_proof(proofid):
+            def invalidate_proof(proofid):
+                self.wait_until(
+                    lambda: can_find_inv_in_poll(
+                        quorum,
+                        proofid,
+                        response=AvalancheProofVoteResponse.REJECTED,
+                        other_response=AvalancheProofVoteResponse.UNKNOWN,
+                    )
+                )
+                return try_rpc(
+                    -8,
+                    "Proof not found",
+                    node.getrawavalancheproof,
+                    uint256_hex(proofid),
+                )
+
+            with node.assert_debug_log(
+                [f"Avalanche invalidated proof {uint256_hex(proofid)}"],
+                ["Failed to reject proof"],
+            ):
+                self.wait_until(lambda: invalidate_proof(proofid))
+
+        # Invalidate the local proof by avalanche vote
+        wait_for_invalidated_proof(proof.proofid)
+
+        assert_equal(
+            node.getavalancheinfo(),
+            {
+                "ready_to_poll": True,
+                "local": {
+                    "verified": False,
+                    "verification_status": "avalanche-invalidated",
+                    "proofid": uint256_hex(proof.proofid),
+                    "limited_proofid": uint256_hex(proof.limited_proofid),
+                    "master": privkey.get_pubkey().get_bytes().hex(),
+                    "stake_amount": coinbase_amount,
+                    "payout_address": ADDRESS_ECREG_UNSPENDABLE,
+                },
+                "network": {
+                    "proof_count": QUORUM_NODE_COUNT,
+                    "connected_proof_count": QUORUM_NODE_COUNT,
+                    "dangling_proof_count": 0,
+                    "finalized_proof_count": 0,
+                    "conflicting_proof_count": 0,
+                    "immature_proof_count": 0,
+                    "total_stake_amount": QUORUM_NODE_COUNT * coinbase_amount,
+                    "connected_stake_amount": QUORUM_NODE_COUNT * coinbase_amount,
+                    "dangling_stake_amount": Decimal("0.00"),
+                    "immature_stake_amount": Decimal("0.00"),
+                    "node_count": QUORUM_NODE_COUNT,
+                    "connected_node_count": QUORUM_NODE_COUNT,
+                    "pending_node_count": 0,
+                },
+            },
+        )
 
 
 if __name__ == "__main__":